+++ flutter_osm_android_jni/lib/src/android_transport/jni_android_map_transport.dart
@@ -2,9 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/services.dart';
 import 'package:flutter_osm_interface/flutter_osm_interface.dart';
-import 'package:flutter_osm_plugin/src/android_jni/generated.dart';
-import 'package:flutter_osm_plugin/src/android_transport/android_event_decoder.dart';
-import 'package:flutter_osm_plugin/src/android_transport/android_map_transport.dart';
+import 'package:flutter_osm_android_jni/src/android_jni/generated.dart';
 import 'package:jni/jni.dart';
 
 /// JNI command transport with MethodChannel acknowledgements and events.